+++ ext/PEtabIpoptExtension.jl
@@ -11,14 +11,14 @@ function PEtab.calibrate_multistart(
         rng::Random.AbstractRNG, prob::PEtabODEProblem, alg::IpoptOptimizer,
         nmultistarts::Integer; nprocs = 1, save_trace = false, dirsave = nothing,
         sample_prior = true, sampling_method = LatinHypercubeSample(),
-        init_weight = nothing, init_bias = nothing, options
+        show_progress = false, init_weight = nothing, init_bias = nothing, options
         ::Union{Nothing, IpoptOptions} = nothing,
     )::PEtab.PEtabMultistartResult
     options = isnothing(options) ? IpoptOptions() : options
 
     return PEtab._calibrate_multistart(
         rng, prob, alg, nmultistarts, dirsave, sampling_method, options, sample_prior,
-        save_trace, nprocs, init_weight, init_bias
+        save_trace, nprocs, show_progress, init_weight, init_bias
     )
 end